rectified newsletter
