+++ src/App.js
@@ -12,7 +12,7 @@ import SignUp from './Components/SignupLogin/SignUp';
 import Login from './Components/SignupLogin/Login';
 import PageNotFound from './Components/Error/PageNotFound';
 import Forget from './Components/Forget/Forget';
-import Newsletter from './Components/Newsletter/newsletter';
+import Newsletter from './Components/HomePage/Sections/Newsletter';
 
 function App() {
   return (
