+++ src/CSS/Newsletter.module.css
@@ -1,90 +1,115 @@
-@media only screen and (min-width: 1200px){
-    .subscribe{
-      display: flex;
-      flex-direction: row;
-    }
-    .subscribe img{
-      width: 700px;
-    }
-    .sub_text{
-        padding: 3rem;
-        margin-top: 6rem;
-        margin-bottom: 3rem;
-        margin-left: 3rem;
-        margin-right: 3rem;
-    }
-    .sub_text fieldset{
-        margin-top: 4%;
-        width: 400px;
-        padding: 5%;
-    }
-  }
-  @media only screen and (min-width: 1000px){
-   
-    .subscribe img{
-      width: 700px;
-    }
-    .sub_text{
-        padding: 3rem;
-        margin-top: 6%;
-        margin-bottom: 3rem;
-        margin-left: 3rem;
-        margin-right: 3rem;
-    }
-    .sub_text fieldset{
-        margin-top: 4%;
-        width: 400px;
-        padding: 5%;
-    }
-  }
   
   
-  @media only screen and (max-width: 1000px){
+  @media (max-width: 420px) {
+    .sub_text fieldset button{
+        font-size: 20px;
+        height: 35px;
+        padding: 2px 12px 2px 12px;
+        border-radius: 4px;
+    }
+     .sub_text fieldset input{
+
+        margin-right: 2rem;
+        margin-bottom: 2rem;
+        height: 40px;
+        width: 200px;
+        border: 1px solid grey;
+        border-radius: 4px;
+    }
     
-    .subscribe img{
-      width: 400px;
+   }
+  
+   @media (min-width: 420px) {
+    .field{
+      margin-left: 0.2rem;
     }
-    .sub_text{
-        padding: 1rem;
-        margin: 6%;
+    .sub_text fieldset button{
+     font-size: 20px;
+     height: 40px;
+     padding: 3px 15px 3px 15px;
+   }
+    .sub_text fieldset input{
+     height: 40px;
+     width: 200px;
+       border: 1px solid grey;
+       border-radius: 4px;
+   }
+   .field{
+     display: flex;
+     flex-direction: row;
+   }
+  }
+   @media (min-width: 992px) {
+     .field{
+       margin-left: 12rem;
+     }
+     .sub_text fieldset button{
+      font-size: 20px;
+      height: 45px;
+      padding: 5px 30px 5px 30px;
     }
-    .sub_text fieldset{
-        margin-top: 4%;
-        width: 300px;
-        padding: 4%;
+     .sub_text fieldset input{
+      height: 45px;
+      width: 280px;
+        border: 1px solid grey;
+        border-radius: 4px;
     }
-  }
-  .sub_text fieldset{
-      border: 1px solid black;
-  }
-  .sub_text fieldset input{
-      padding: 10px;
-      margin: 10px;
-      width: 200px;
-      height: 30px;
-      border: 1px solid grey;
-      border-radius: 4px;
-  }
-  .name , .email{
+    .field{
       display: flex;
       flex-direction: row;
+    }
+   }
+   @media (min-width: 1200px) {
+    .field{
+      margin-left: 17rem;
+    }
+    .field{
+      display: flex;
+      flex-direction: row;
+    }
+    .sub_text fieldset input{
+      height: 45px;
+      width: 300px;
+        border: 1px solid grey;
+        border-radius: 4px;
+    }
+    .sub_text fieldset button{
+      padding: 5px 30px 5px 30px;
+      font-size: 20px;
+      height: 45px;
+    }
+   }
+
+  .subscribe{
+    margin-left: 10%;
+    padding: 7%;
   }
+  .sub_text h1{
+    color: '#008dc8';
+    font-weight: bold;
+    font-size: 40px;
+    font-family: 'Poppins', sans-serif;
+    letter-spacing: 0px;
+    text-align: center;
+  }
+  .sub_text h4{
+    font-family: 'Poppins', sans-serif;
+    letter-spacing: 0px;
+    text-align: center;
+  }
+ 
+ 
+  .sub_text fieldset{
+      padding: 2rem;  
+  }
+  
   .sub_text fieldset button{
     float: right;
-    margin-top: 12px;
   margin-right: 10%;
-  padding: 5px 30px 5px 30px;
   color: #fff;
-  font-size: 20px;
   text-align: center;
   background-color: #008dc8;
   border: none;
-  border-radius: 50px;
   }
-  .span_email , .span_name{
-      width: 30px;
-      height: 30px;
-      padding: 3px;
-      background-color: #008dc8;
-      color: white;
-  }
+ 
+
